New export `css-doodle/component` for importng CSSDoodle class itself
