+++ src/exports/component/index.js
@@ -1,0 +1,5 @@
+import { CSSDoodle, register } from '../../component.js';
+
+export {
+  CSSDoodle, register
+}
